Migrate to null-safety
For this we need to bump the minimum supported Dart version to 2.12 and
upgrade the depepndencies to *nullsafety* pre-releases.

The `coverage` package doesn't really need to be null-safe because we
only use it to run a tool from it (and when generating coverage reports
only), which works fine with a null-safety ready dart compiler anyway.