workflow/wip: Don't check draft pull requests
If a PR is marked as a draft, then we know there might be WIPs, and that
is fine. Making the CI red doesn't help the PR author to know if his
working is doing well or not. Once the PR is marked as ready for review
(for which now the workflow also run for that action), we should make
sure there are no WIP commits left and the PR is not marked as WIP
anymore either.