Make wip action really fail when there is a WIP commit
When there are WIP commits, the error code is not properly set because:

```sh
  echo "@wips" | while read wip
```

makes the `while` loop run in a new shell, so when the `r=$(($r+1))` is
done is not propagated to the parent shell (the one running the main
script) so the `r` in it is not updated, ending up with an `exit 0`
always.

We just run the `while` loop in the main shell and use bash heredocs to
pipe the input in.