ci: Remove minor version from tested version
This way we make sure we test using the latest dart version. At the
moment only the latest version is supported because null-safety support
can only be enabled starting from 2.12, in the future all the oldest
dart versions that should be supported have to be added to the matrix
explicitly.

Signed-off-by: Leandro Lucarella <luca@llucax.com>